requirements.txt
python-telegram-bot==20.7
anthropic==0.18.1
python-dotenv==1.0.1
aiohttp==3.9.3
requests==2.31.0 